Homework 3 Changes

+++ homework3/src/Task1.java
@@ -3,8 +3,9 @@ public class Task1 {
         int x = 47;
         int y = 33;
         int z = 10;
-        System.out.println(getAverage(x, y, z));
+        System.out.println((double) getAverage(x, y, z));
     }
+
     static int getAverage(int x, int y, int z){
         return x > z ? (x + y) / 2 : z;
     }
